pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finxcloud"
version = "0.2.0"
description = "Multi-Cloud Cost Optimization Tool — FinXCloud (AWS, Azure, GCP)"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.15.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-sql>=4.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-containerservice>=28.0.0",
    "azure-mgmt-resource>=23.0.0",
]
gcp = [
    "google-auth>=2.28.0",
    "google-cloud-compute>=1.16.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-container>=2.40.0",
    "google-cloud-billing>=1.13.0",
    "google-cloud-bigquery>=3.17.0",
    "google-cloud-resource-manager>=1.12.0",
    "google-api-python-client>=2.120.0",
]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]
pdf = [
    "reportlab>=4.1.0",
]
all-clouds = [
    "finxcloud[azure,gcp]",
]
dev = [
    "pytest>=7.0",
    "hypothesis>=6.0",
    "moto>=5.0",
    "ruff>=0.3.0",
]

[project.scripts]
finxcloud = "finxcloud.cli:main"

[tool.setuptools.packages.find]
include = ["finxcloud*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["finxcloud"]

[tool.setuptools.package-data]
finxcloud = ["py.typed"]
